style(Lotto): move the generateTickets funcion to Lotto companion class, make InputView and OutputView cleaner
